Use a my.cnf file in the client for username/password
Putting the password in the command-line code throws mysql errors. This
commit takes out the username and password and passes in a .my.cnf file.
The expectation is that clients of this code will have that file in a
config directory.

If this code was used by more than one client, I'd make this
configurable. But it's only used by dw_import_scripts2 so I'm fine
making a hard-coded dependency like this.