Backport the TableNamer module
The dw_import_scripts2 project uses this library and it's hardlocked to
a version 1.0.0. I've been testing it with 1.1.0 and it mostly works.
But there are some calls to TableNamer, which wasn't available in the
version of OTM it was supposed to be using.

How this worked in production, I'm not sure.

i'm trying to make the smallest possible change by back porting the
TableNamer code into this 1.1.0-based branch.