Standardize on [table_name]_old for archive tables
Between the two versions of oracle_to_mysql (and some other code) we
ended up with two ways of naming archive tables. Version one is

[table_name]_old

And version two is

old_[date time]_[table_name]

The goal of the 2nd form was to allow us to store multiple days of
archives. But, as far as I can see, there is no actual use for multiple
days of archives. They just take up space.

And having table names in the 2nd form causes havoc for granting access.
In MySQL you can grant a user access to a specific table or all tables.
You can not grant them access to a table name with a wild card in it. So
if, say, student athletes needs to select data from ps_person and
compare it against yesterday's archive, then the first form makes this
easy:

grant select to 'ps_person'
grant select to 'ps_person_old'

But the 2nd form offers no easy way to do that. You can't do

grant select to 'old_%_ps_person'

So you either have to give the user select access to all tables. Or you
have to programatically re-assign grants each night.

If we were using the multi-day archive then this might (might) be worth
the hassle. But we're not, so it's not, so I'm removing it.

This commit will require that DW_Import_Scripts be updated as well,
since it names the archive tables inconsistently.